requirements.txt
# Core
streamlit==1.26.1
PyPDF2==3.0.1
faiss-cpu==1.7.4

# LangChain + Google GenAI
langchain==0.1.16
langchain-google-genai==0.0.5


# Pydantic (v1 to avoid v2 issues)
pydantic==1.10.14

# Optional utilities
joblib==1.3.2
python-dotenv==1.0.1

ipyKernel


